# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Birrias API",
        "description": "API RESTful completa para gestión de torneos de fútbol amateur desarrollada con Laravel 12 + Breeze API + Sanctum",
        "contact": {
            "name": "Birrias API Support",
            "email": "admin@birrias.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Birrias API Server"
        }
    ],
    "paths": {
...
        "/api/standings": {
            "get": {
                "tags": [
                    "Standings"
                ],
                "summary": "Listar tablas de posiciones",
                "description": "Obtiene las tablas de posiciones con filtros opcionales",
                "operationId": "f66b99401dd56a836f385ae41a2b2d08",
                "parameters": [
                    {
                        "name": "tournament_id",
                        "in": "query",
                        "description": "Filtrar por torneo",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "Filtrar por grupo",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de posiciones ordenada por puntos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "tournament_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "team_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "group_id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "nullable": true
                                            },
                                            "matches_played": {
                                                "type": "integer"
                                            },
                                            "wins": {
                                                "type": "integer"
                                            },
                                            "draws": {
                                                "type": "integer"
                                            },
                                            "losses": {
                                                "type": "integer"
                                            },
                                            "goals_for": {
                                                "type": "integer"
                                            },
                                            "goals_against": {
                                                "type": "integer"
                                            },
                                            "goal_difference": {
                                                "type": "integer"
                                            },
                                            "points": {
                                                "type": "integer"
                                            },
                                            "tournament": {
                                                "type": "object"
                                            },
                                            "team": {
                                                "type": "object"
                                            },
                                            "group": {
                                                "type": "object",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/standings/{id}": {
            "get": {
                "tags": [
                    "Standings"
                ],
                "summary": "Ver posición específica",
                "description": "Obtiene la información detallada de una posición específica en la tabla",
                "operationId": "b3e8d9d7f12e2445f4606839c550bf63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la posición",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información de la posición",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "tournament_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "team_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "group_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "nullable": true
                                        },
                                        "matches_played": {
                                            "type": "integer"
                                        },
                                        "wins": {
                                            "type": "integer"
                                        },
                                        "draws": {
                                            "type": "integer"
                                        },
                                        "losses": {
                                            "type": "integer"
                                        },
                                        "goals_for": {
                                            "type": "integer"
                                        },
                                        "goals_against": {
                                            "type": "integer"
                                        },
                                        "goal_difference": {
                                            "type": "integer"
                                        },
                                        "points": {
                                            "type": "integer"
                                        },
                                        "tournament": {
                                            "type": "object"
                                        },
                                        "team": {
                                            "type": "object"
                                        },
                                        "group": {
                                            "type": "object",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Posición no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/standings/tournament/{tournamentId}": {
            "get": {
                "tags": [
                    "Standings"
                ],
                "summary": "Tabla de posiciones por torneo",
                "description": "Obtiene la tabla de posiciones de un torneo específico",
                "operationId": "ad2fe6aab3903a5d850dcb0423fc8017",
                "parameters": [
                    {
                        "name": "tournamentId",
                        "in": "path",
                        "description": "ID del torneo",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tabla de posiciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/standings/group/{groupId}": {
            "get": {
                "tags": [
                    "Standings"
                ],
                "summary": "Tabla de posiciones por grupo",
                "description": "Obtiene la tabla de posiciones de un grupo específico",
                "operationId": "f8f5c07491fd6d1cd0c36f174d52f032",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "ID del grupo",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tabla de posiciones del grupo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "tournament_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "team_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "group_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "matches_played": {
                                                "type": "integer"
                                            },
                                            "wins": {
                                                "type": "integer"
                                            },
                                            "draws": {
                                                "type": "integer"
                                            },
                                            "losses": {
                                                "type": "integer"
                                            },
                                            "goals_for": {
                                                "type": "integer"
                                            },
                                            "goals_against": {
                                                "type": "integer"
                                            },
                                            "goal_difference": {
                                                "type": "integer"
                                            },
                                            "points": {
                                                "type": "integer"
                                            },
                                            "team": {
                                                "type": "object"
                                            },
                                            "tournament": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Grupo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}